fix for form underlapping nav bar
Fixed an issue where in certain circumstances, the login form in
SPLoginLogicViewController would underlap the navigation bar. This
issue can be reproduced by attempting to use this login controller
running an iPhone-native app on an iPad. This fix prevents the edges
from being extended, ensuring that the login form never gets
underlapped.